Adjusted game loop

+++ Lebakas/src/Lebakas/Application.h
@@ -14,7 +14,7 @@ namespace Lebakas
 		void Run();
 		void ProcessInput();
 
-		virtual void Update(float deltaTime);
+		virtual void Update(double deltaTime);
 
 	protected:
 		WindowManager mRenderer;
